Move batchers into their own module (and registry)
